Starts implementing GoApi
